+++ save-api/src/main/kotlin/com/saveourtool/save/api/SaveCloudClient.kt
@@ -12,10 +12,10 @@ import com.saveourtool.save.api.utils.submitExecution
 import com.saveourtool.save.api.utils.uploadAdditionalFile
 import com.saveourtool.save.domain.ProjectCoordinates
 import com.saveourtool.save.domain.ShortFileInfo
-import com.saveourtool.save.entities.RunExecutionRequest
 import com.saveourtool.save.execution.ExecutionDto
 import com.saveourtool.save.execution.ExecutionStatus
 import com.saveourtool.save.execution.TestingType
+import com.saveourtool.save.request.CreateExecutionRequest
 import com.saveourtool.save.utils.DATABASE_DELIMITER
 
 import arrow.core.Either
@@ -103,8 +103,8 @@ class SaveCloudClient(
     private suspend fun submitExecution(
         additionalFiles: List<ShortFileInfo>?,
         contestName: String?,
-    ): Either<HttpStatusCode, RunExecutionRequest> {
-        val runExecutionRequest = RunExecutionRequest(
+    ): Either<HttpStatusCode, CreateExecutionRequest> {
+        val createExecutionRequest = CreateExecutionRequest(
             projectCoordinates = ProjectCoordinates(
                 organizationName = evaluatedToolProperties.organizationName,
                 projectName = evaluatedToolProperties.projectName,
@@ -119,34 +119,34 @@ class SaveCloudClient(
             testingType = testingType,
             contestName = contestName,
         )
-        val response = httpClient.submitExecution(runExecutionRequest)
+        val response = httpClient.submitExecution(createExecutionRequest)
         val httpStatus = response.status
         if (httpStatus !in arrayOf(OK, Accepted)) {
-            log.error("Received HTTP $httpStatus while submitting execution: $runExecutionRequest")
+            log.error("Received HTTP $httpStatus while submitting execution: $createExecutionRequest")
             val responseBody = response.bodyAsText()
             if (responseBody.isNotBlank()) {
                 log.error("HTTP response body: $responseBody")
             }
             return httpStatus.left()
         }
-        return runExecutionRequest.right()
+        return createExecutionRequest.right()
     }
 
     /**
-     * Get results for current [runExecutionRequest]:
+     * Get results for current [createExecutionRequest]:
      * sending requests, which checks current state of execution, until it will be finished, or timeout will be reached
      *
-     * @param runExecutionRequest
+     * @param createExecutionRequest
      */
     @Suppress("MagicNumber")
     private suspend fun getExecutionResults(
-        runExecutionRequest: RunExecutionRequest,
+        createExecutionRequest: CreateExecutionRequest,
     ): ExecutionDto? {
         // Execution should be processed in db after submission, so wait little time
         delay(1_000)
 
         // We suppose, that in this short time (after submission), there weren't any new executions, so we can take the latest one
-        val executionId = httpClient.getLatestExecution(runExecutionRequest.projectCoordinates.projectName, runExecutionRequest.projectCoordinates.organizationName).id
+        val executionId = httpClient.getLatestExecution(createExecutionRequest.projectCoordinates.projectName, createExecutionRequest.projectCoordinates.organizationName).id
 
         var executionDto = httpClient.getExecutionById(executionId)
         val initialTime = LocalDateTime.now()
